refactor: HomeFragment DataBinding 적용
