+++ app/src/androidTest/java/woowacourse/movie/HomeFragmentTest.kt
@@ -15,7 +15,7 @@ import org.hamcrest.Matchers.allOf
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.movie.activity.MainActivity
+import woowacourse.movie.activity.main.MainActivity
 import woowacourse.movie.movie.MovieMockData
 
 class HomeFragmentTest {
